update cl-dice based losses cont.
